backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
openai==1.12.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
